Set the release profile as default
Trying to run the jar produced by the debug profile with "java -jar" will result in a message "Could not find or load main class com.networknt.server.Server". This prevents the light4j service in the container from running and leads to the automated tests to report a RegistryApplicationNotFoundException.

The reason for editing the pom is that it does not seem like there is any way to specify the profile elsewhere in the test execution process.